[Refactor] Fix BE UT on MacOS

+++ be/test/exec/tablet_sink_index_channel_test.cpp
@@ -323,7 +323,7 @@ void TabletSinkIndexChannelTest::test_load_diagnose_base(const std::string& erro
     auto tuple_desc = runtime_state->desc_tbl().get_tuple_descriptor(_desc_tbl.tupleDescriptors[0].id);
     ChunkUniquePtr chunk = ChunkHelper::new_chunk(*tuple_desc, 1);
     chunk->get_column_raw_ptr_by_index(0)->append_datum(Datum(1));
-    chunk->get_column_raw_ptr_by_index(1)->append_datum(Datum(1L));
+    chunk->get_column_raw_ptr_by_index(1)->append_datum(Datum(int64_t(1)));
     ASSERT_OK(sink->send_chunk(runtime_state.get(), chunk.get()));
     ASSERT_FALSE(sink->close(runtime_state.get(), Status::OK()).ok());
     ASSERT_EQ(expected_num_stack_trace, num_stack_trace);
@@ -471,7 +471,7 @@ TEST_F(TabletSinkIndexChannelTest, send_request_releases_protobuf_memory) {
     auto tuple_desc = runtime_state->desc_tbl().get_tuple_descriptor(_desc_tbl.tupleDescriptors[0].id);
     ChunkUniquePtr chunk = ChunkHelper::new_chunk(*tuple_desc, 1);
     chunk->get_column_raw_ptr_by_index(0)->append_datum(Datum(1));
-    chunk->get_column_raw_ptr_by_index(1)->append_datum(Datum(1L));
+    chunk->get_column_raw_ptr_by_index(1)->append_datum(Datum(int64_t(1)));
     ASSERT_OK(sink->send_chunk(runtime_state.get(), chunk.get()));
     // close() flushes the buffered chunk and triggers _send_request with eos=true,
     // which serializes the chunk data and dispatches the RPC.
